minor UI overhaul; added devcap functionality;
This is some pretty big stuff. Now that you are able to have more than
one developer, you are able to work on more than one project at a time,
thus getting more stuff done. Needs to divide the remaining # of days
between the # of devs on your team, because there is no way anyone is
going to click through the 150+ days of level 8-10;